requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
ansible==6.5.0
ansible-compat==2.2.1
ansible-core==2.13.5
ansible-lint==6.8.2
astroid==2.12.13
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
bracex==2.3.post1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cryptography==38.0.1
dill==0.3.6
discord.py==2.1.0
docopt==0.6.2
filelock==3.8.0
frozenlist==1.3.3
idna==3.4
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
jsonschema==4.16.0
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
mutagen==1.46.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
pep8==1.7.1
pipreqs==0.4.11
platformdirs==2.5.2
pluggy==1.0.0
powerline-status==2.7
py==1.11.0
pycparser==2.21
pycryptodomex==3.15.0
Pygments==2.13.0
pylint==2.15.7
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
PyYAML==6.0
requests==2.28.1
resolvelib==0.8.1
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
subprocess-tee==0.3.5
tomli==2.0.1
tomlkit==0.11.6
urllib3==1.26.13
wcmatch==8.4.1
websockets==10.3
wrapt==1.14.1
yamllint==1.28.0
yarg==0.1.9
yarl==1.8.1
yt-dlp==2022.10.4
